[Refactor] 순환참조 해결(#16)
