+++ planfit/src/main/java/com/planfit/server/service/RoutineService.java
@@ -5,10 +5,14 @@ import com.planfit.server.common.message.ErrorMessage;
 import com.planfit.server.domain.Exercise;
 import com.planfit.server.domain.Routine;
 import com.planfit.server.domain.User;
+import com.planfit.server.dto.request.ExerciseReorderRequest;
+import com.planfit.server.dto.response.ExerciseGetAllResponse;
 import com.planfit.server.repository.RoutineRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Service
 @Transactional(readOnly = true)
@@ -67,6 +71,30 @@ public class RoutineService {
         return routineRepository.findByUserAndExercise(user, exercise).orElseThrow(
                 () -> new NotFoundException(ErrorMessage.ROUTINE_NOT_FOUND)
         );
+    }
+
+    //운동 리스트 조회
+    public ExerciseGetAllResponse findRoutines(Long userId) {
+        User user = userService.getUserById(userId);
+
+        return ExerciseGetAllResponse.fromRoutines(routineRepository.findAllByUserOrderBySequenceAsc(user));
+    }
+
+    @Transactional
+    //운동 리스트 순서 변경
+    public void reorderExercises(Long userId, List<ExerciseReorderRequest> exercises) {
+        User user = userService.getUserById(userId);
+
+        //새로운 순서로 업데이트
+        updateRoutinesWithNewOrder(exercises, user);
+    }
+
+    private void updateRoutinesWithNewOrder(List<ExerciseReorderRequest> exercises, User user) {
+        exercises.stream().forEach(request -> {
+            Exercise exercise = exerciseService.getExerciseById(request.id());
+            Routine routine = getRoutineByUserAndExercise(user, exercise);
+            routine.updateSequence(request.index());
+        });
     }
 
 }
